Basic structure for networking
I have implemented a basic structure for network with some traits,
components, structs, etc. The actualy logic for the NetworkServerPlugin
is not finished yet but will implement it soon.

I am trying to get a good networking base going so that future
development of networking related stuff is easy and it is only a matter
of creating packets or adding some components or traits to some stuff.

Once basic client <-> server networking is done I will probably go onto
server <-> server communication as this is most likely the most
difficuly and complex part of this entire project. Implementing the
networking communication between servers is very important because if
that fails then I will most likely have to go back and reimplement a lot
of the networking backend which I do not want to do. Please bare with me
while I attempt to create a decent networking protocol using QUIC and
quinn library.

And to who ever might stumble upon this and have some technical
expertise with networking, it would be greatly appreciated for any
feedback or suggestions.

Thank you and have a great day!